+++ MuPDF.NET/Barcode/Encoders/Barcode.cs
@@ -1,10 +1,4 @@
-﻿/**************************************************
- Bytescout BarCode SDK
- Copyright (c) 2008 - 2010 Bytescout
- All rights reserved
- www.bytescout.com
-**************************************************/
-
+﻿
 using System;
 using System.Collections.Generic;
 using System.Globalization;