+++ MuPDF.NET/Barcode/Encoders/BarcodeException.cs
@@ -1,9 +1,3 @@
-/**************************************************
- Bytescout BarCode SDK
- Copyright (c) 2008 - 2010 Bytescout
- All rights reserved
- www.bytescout.com
-**************************************************/
 
 using System;
 using System.Text;